backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg[binary]>=3.2.0
sqlalchemy>=2.0.35
python-dotenv>=1.0.1
pydantic>=2.9.0
pandas>=2.2.0
scikit-learn>=1.5.0
requests>=2.32.0
